use autoretry on failure feature of upcoming Behave 1.2.6
As Behave 1.2.6 is not released yet I had to install it directly from git repo.
I've pinned it to the last commit ID available today on master branch.
Thanks to this failing scenarios will be automatically re-run for the second
time. If it fails twice in a row, then scenarios will be marked as failed.
This approach should allow us to reduce the number of tests failing due to
slowly loading pages or some intermittent network slow-downs.